requirements.txt
fastapi
uvicorn
yfinance
pandas
httpx
python-dotenv
pydantic
